# doc/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Chista | Open Source Threat Intelligence Framework",
    "description": "Chista is an Open Source Cyber Threat Intelligence (CTI) Framework designed to help users understand, predict and defend against cyber threats. You can find out more about\nChista at [github repository](https://github.com/ChistaDev/Chista). You can now help us improve the API whether it's by making changes to the definition itself or to the code.\nThat way, with time, we can improve the API in general, and expose some of the new features.\n\nSome useful links:\n- [Chista Repository](https://github.com/ChistaDev/Chista)\n- [Chista Documentation](https://github.com/ChistaDev/Chista/blob/main/chista-docs-v1.md)",
    "contact": {
      "email": "info@chista.github.io"
    },
    "license": {
      "name": "MIT License",
      "url": "https://github.com/ChistaDev/Chista/blob/main/LICENSE"
    },
    "version": "1.0.2"
  },
  "servers": [
    {
      "url": "https://localhost:7777/api/v1"
    }
  ],
  "paths": {
    "/blacklist": {
      "summary": "provide information about the blacklist status of a given IP or domain",
      "description": "The Blacklist API is designed to provide information about the blacklist       status of a given IP or domain. It interacts with the mxtoolbox.com       website to fetch and analyze the blacklist data.The API uses chromedp to       interact with the mxtoolbox.com website. The HTML response is scraped to extract relevant information from the blacklist table.",
      "get": {
        "summary": "",
        "description": "It interacts with the mxtoolbox.com website to fetch and analyze the blacklist data.",
        "operationId": "checkBlacklist",
        "parameters": [
          {
            "name": "asset",
            "in": "query",
            "description": "The IP address or domain to be checked",
            "required": true,
            "schema": {
              "type": "string",
              "example": "example.com"
            }
          },
          {
            "name": "verbosity",
            "in": "query",
            "description": "Specifies the verbosity level for logging (Optional)",
            "required": false,
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "HTTP 200 OK",
            "content": {
              "application/json": {
                "examples": {
                  "Blacklisted": {
                    "value": "[\n    {\n        \"status\": \"Status Blacklisted\",\n        \"name\": \"Spamhaus ZEN\",\n        \"link\": \"https://mxtoolbox.com/Problem/Blacklist/Spamhaus-ZEN/?page=prob_blacklist&amp;ip=61.53.91.245&amp;link=button&amp;action=blacklist:61.53.91.245&amp;showLogin=1&amp;hidetoc=1&amp;reason=127.0.0.11, 127.0.0.4\"\n    },\n    {\n        \"status\": \"Status Blacklisted\",\n        \"name\": \"UCEPROTECTL3\",\n        \"link\": \"https://mxtoolbox.com/Problem/Blacklist/UCEPROTECTL3/?page=prob_blacklist&amp;ip=61.53.91.245&amp;link=button&amp;action=blacklist:61.53.91.245&amp;showLogin=1&amp;hidetoc=1&amp;reason=127.0.0.2\"\n    }\n]"
                  },
                  "Not blacklisted": {
                    "value": "{\"message\":\"IP/Domain is not blacklisted.\"}"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "blacklist"
        ]
      }
    },
...
  },
...
}